[#2434] Make packages in application modules follow the pattern in 'clients'.
